Only download database assets in development environment if they do not exist
Any image built in CI should not have these assets present on filesystem
so should always download fresh up-to-date version

+++ Makefile
@@ -18,7 +18,9 @@ sync:
 deploy: databases push release
 
 databases::
-	curl -qsfL -o digital-land.sqlite3 https://digital-land-production-collection-dataset.s3.eu-west-2.amazonaws.com/digital-land-builder/dataset/digital-land.sqlite3
+# Prefer local sqlite3 if available (e.g. in development environment to reduce rebuild time)
+# When run in CI this should always pull a fresh version when building image
+	[ -f digital-land.sqlite3 ] || curl -qsfL -o digital-land.sqlite3 https://digital-land-production-collection-dataset.s3.eu-west-2.amazonaws.com/digital-land-builder/dataset/digital-land.sqlite3
 	bin/load.sh
 
 black:
